refactor: profile sync controller - add multiple file exports
