+++ packages/profile-sync-controller/package.json
@@ -22,6 +22,31 @@
       "require": "./dist/index.js",
       "types": "./dist/types/index.d.ts"
     },
+    "./sdk": {
+      "import": "./dist/sdk/index.mjs",
+      "require": "./dist/sdk/index.js",
+      "types": "./dist/types/sdk/index.d.ts"
+    },
+    "./auth": {
+      "import": "./dist/controllers/authentication/index.mjs",
+      "require": "./dist/controllers/authentication/index.js",
+      "types": "./dist/types/controllers/authentication/index.d.ts"
+    },
+    "./user-storage": {
+      "import": "./dist/controllers/user-storage/index.mjs",
+      "require": "./dist/controllers/user-storage/index.js",
+      "types": "./dist/types/controllers/user-storage/index.d.ts"
+    },
+    "./dist/mocks/auth": {
+      "import": "./dist/controllers/authentication/__fixtures__/index.mjs",
+      "require": "./dist/controllers/authentication/__fixtures__/index.js",
+      "types": "./dist/types/controllers/authentication/__fixtures__/index.d.ts"
+    },
+    "./dist/mocks/user-storage": {
+      "import": "./dist/controllers/user-storage/__fixtures__/index.mjs",
+      "require": "./dist/controllers/user-storage/__fixtures__/index.js",
+      "types": "./dist/types/controllers/user-storage/__fixtures__/index.d.ts"
+    },
     "./package.json": "./package.json"
   },
   "main": "./dist/index.js",
