+++ packages/profile-sync-controller/src/controllers/authentication/index.ts
@@ -1,5 +1,6 @@
 import Controller from './AuthenticationController';
 
+const AuthenticationController = Controller;
 export { Controller };
+export default AuthenticationController;
 export * from './AuthenticationController';
-export * as Mocks from './__fixtures__';
