+++ packages/profile-sync-controller/src/controllers/user-storage/index.ts
@@ -1,6 +1,7 @@
 import Controller from './UserStorageController';
 
+const UserStorageController = Controller;
 export { Controller };
+export default UserStorageController;
 export * from './UserStorageController';
 export * from './encryption';
-export * as Mocks from './__fixtures__';
